+++ src/Services/CryptoService.cs
@@ -1,4 +1,4 @@
-using XPoster.Abstraction;
+using XPoster.Contracts;
 
 namespace XPoster.Services;